Unify key use in app
