+++ lib/widgets/header/header.dart
@@ -1,34 +1,24 @@
 import 'package:flutter/material.dart';
 import 'package:responsive_builder/responsive_builder.dart';
 import 'package:seo_renderer/seo_renderer.dart';
-import 'package:wolnakeja/home/homeview.dart';
 import 'package:wolnakeja/styles/app_colors.dart';
 import 'package:wolnakeja/styles/app_text_styles.dart';
+import 'package:wolnakeja/widgets/app_navigation_bar.dart';
 import 'package:wolnakeja/widgets/dividerek.dart';
-import 'package:wolnakeja/widgets/navigation_bar.dart';
 
 part 'header_mobile.dart';
 part 'header_tablet.dart';
 part 'header_desktop.dart';
 
 class Header extends StatelessWidget {
-  const Header({
-    super.key,
-    // required this.navigationItemsKeys,
-  });
-
-  //final List<GlobalKey> navigationItemsKeys;
+  const Header({super.key});
 
   @override
   Widget build(BuildContext context) {
     return ScreenTypeLayout(
       mobile: const HeaderMobile(),
-      tablet: HeaderTablet(
-        navigationItemsKeys: navigationItemsKeys,
-      ),
-      desktop: HeaderDesktop(
-        navigationItemsKeys: navigationItemsKeys,
-      ),
+      tablet: const HeaderTablet(),
+      desktop: const HeaderDesktop(),
     );
   }
 }
