+++ pubspec.yaml
@@ -13,7 +13,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   flutter:
     sdk: flutter
-  responsive_builder: ^0.4.1
+  responsive_builder: ^0.6.0
   seo_renderer: ^0.4.0
   spaced: ^1.0.3
   url_launcher: ^6.0.20
